test: implement simple ssh server
The ssh server is not fully functional and includes functionality only
suitable for running as part of this test suite.

Also remove the dependency on buffer-equal-constant-time that somehow
snuck in. If this was meant to be a real server we would care about
constant-time credential verification, but since this server is
explicitly for testing, it just represents an extra dependency that
isn't really needed.

deps: remove unecessary buffer comparison